+++ Runtime/InstantiateGeyserPrefab.cs
@@ -2,6 +2,7 @@
 using UnityEngine;
 using UnityEngine.Networking;
 using RoR2;
+using RoR2.Navigation;
 
 namespace LOP
 {
@@ -13,27 +14,37 @@ namespace LOP
     
     public class InstantiateGeyserPrefab : MonoBehaviour
     {
-        private enum GeyserType
+        public enum GeyserType
         {
             Default = 0,
             Ambry = 1,
             Moon = 2,
             Aphelian = 3,
             Siphoned = 4,
-            Void = 5
+            Void = 5,
+            Fan = 6
         }
 
         [Tooltip("The prefab used for the geyser")]
-        [SerializeField] private GeyserType geyserType;
+        [SerializeField] public GeyserType geyserType;
         [Tooltip("When the prefab is instantiated, and this is true, the prefab's position and rotation will be set to 0")]
-        [SerializeField] private bool setPositionAndRotationToZero;
+        [SerializeField] private bool setPositionAndRotationToZero = true;
         [Tooltip("setPositionAndRotationToZero would work relative to it's parent")]
-        [SerializeField] private bool useLocalPositionAndRotation;
+        [SerializeField] private bool useLocalPositionAndRotation = true;
         [Tooltip("Wether the Refresh method will be called in the editor")]
-        [SerializeField] private bool refreshInEditor;
+        [SerializeField] private bool refreshInEditor = true;
         [Tooltip("Wether the sound string remains the same as the original instance")]
         [SerializeField] private bool conserveSoundString;
+
+        //Fan specific
+        [HideInInspector] public bool gateToggleOnPurchase;
+        [HideInInspector] public string gateToEnableWhenPurchased;
+        [HideInInspector] public string gateToDisableWhenPurchased;
+
         [SerializeField, HideInInspector] private bool hasNetworkIdentity;
+
+        
+
 
 
         private string address;
@@ -85,6 +96,9 @@ namespace LOP
                 case 5:
                     address = "RoR2/DLC1/voidstage/mdlVoidGravityGeyser.prefab";
                     break;
+                case 6:
+                    address = "RoR2/Base/frozenwall/HumanFan.prefab";
+                    break;
                 default:
                     LOPLog.Error($"This isn't supposed to print in {this}. Geyser Type is invalid.");
                     return;
@@ -94,13 +108,32 @@ namespace LOP
             instance = Instantiate(prefab, transform);
 
             JumpVolume jVolOG = gameObject.GetComponent<JumpVolume>();
-            JumpVolume jVolIn = instance.GetComponentInChildren<JumpVolume>();
-            jVolIn.targetElevationTransform = jVolOG.targetElevationTransform;
+            JumpVolume jVolIn = instance.GetComponentInChildren<JumpVolume>(true);
+
+            if (jVolOG.targetElevationTransform)
+                jVolIn.targetElevationTransform = jVolOG.targetElevationTransform;
+            else
+                LOPLog.Warning($"Target elevation transform on {this} is invalid.");
+
             jVolIn.jumpVelocity = jVolOG.jumpVelocity;
+
             jVolIn.time = jVolOG.time;
-            if(!conserveSoundString)
+            if (!conserveSoundString)
+            {
+                if(String.IsNullOrEmpty(jVolOG.jumpSoundString))
+                    LOPLog.Warning($"Jump sound string is empty on {this}. This will result in silence for your geyser.");
                 jVolIn.jumpSoundString = jVolOG.jumpSoundString;
+            }
+                
             jVolIn.onJump = jVolOG.onJump;
+
+            if(geyserType == GeyserType.Fan && gateToggleOnPurchase)
+            {
+                GateStateSetter gateStateSetter = jVolIn.gameObject.AddComponent<GateStateSetter>();
+                gateStateSetter.gateToEnableWhenEnabled = gateToEnableWhenPurchased;
+                gateStateSetter.gateToDisableWhenEnabled = gateToDisableWhenPurchased;
+            }
+
             jVolOG.enabled = false;
 
             instance.hideFlags |= HideFlags.DontSaveInEditor | HideFlags.DontSaveInBuild | HideFlags.NotEditable;