+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+**1.1.0**
+- Added new GameObject Menu item called ``Risk of Rain 2``.
+- New ``Geyser`` prefab option under ``Risk of Rain 2`` menu Item. Used to quickly make custom jump pads.
+- Added ``InstantiateGeyserPrefab`` for reusing in game geysers.
+- Added ``InstantiateLogbookPrefab`` for instantiating the logbook pickup usually found on hidden realms.
+- Made ``Set Position And Rotation to Zero``, ``Use Local Position and Rotation``, and ``Refresh in Editor`` all set to true by default.
+
+
 **1.0.0**
 
 - Initial Release!